Add verification scripts for dynamodb backup
After backup script is finished, a verification is needed to ensure the records were sucessfully backed up.
This script will pull a record from athena and transform the data to a python dictionary. Based on the table
key, a query will be made to dynamodb and the two records will be compared to ensure data integrity.